Puzzle parts finished
Just need to make sure all the parts are properly linked and and debugged.
If there is extra time, I'd like to make a cube spawner and require boxes sit on the pressure pads, and disappear when the puzzle is reset.
Buttons need to be updated as Basic Button abstract class cannot be used on buttons.